+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.9.7"></a>[v1.9.7] - 2023-10-17
+
 ### Added
 
 - Slovak translation (thanks Martin Frandel, Milan Šalka)
@@ -19,6 +21,10 @@ All notable changes to this project will be documented in this file.
 ### Removed
 
 - Map: Stamen Toner layer (no longer served for free by Stamen)
+
+### Fixed
+
+- crash when playing video on devices with hardened malloc
 
 ## <a id="v1.9.6"></a>[v1.9.6] - 2023-09-25