+++ CHANGELOG.md
@@ -4,6 +4,28 @@ All notable changes to this project will be documented in this file.
 
 ## <a id="unreleased"></a>[Unreleased]
 
+## <a id="v1.9.7"></a>[v1.9.7] - 2023-10-17
+
+### Added
+
+- Slovak translation (thanks Martin Frandel, Milan Šalka)
+- Vietnamese translation (thanks ngocanhtve, Le Nhut Binh)
+
+### Changed
+
+- mosaic layout: clamp ratio to 32/9
+- Video: disable subtitles by default
+- Map: Stamen Watercolor layer (no longer served for free by Stamen) now served by Smithsonian Institution
+- upgraded Flutter to stable v3.13.7
+
+### Removed
+
+- Map: Stamen Toner layer (no longer served for free by Stamen)
+
+### Fixed
+
+- crash when playing video on devices with hardened malloc
+
 ## <a id="v1.9.6"></a>[v1.9.6] - 2023-09-25
 
 ### Fixed